initial file structure cleanup
