+++ woodwork/deserializers/arrow_deserializer.py
@@ -1,0 +1,5 @@
+from woodwork.deserializers import FeatherDeserializer
+
+
+class ArrowDeserializer(FeatherDeserializer):
+    format = "arrow"
