+++ woodwork/deserializers/pickle_deserializer.py
@@ -1,0 +1,9 @@
+from woodwork.deserializers.deserializer_base import Deserializer
+
+
+class PickleDeserializer(Deserializer):
+    format = "pickle"
+
+    def read_from_local_path(self):
+        lib = self._get_library()
+        return lib.read_pickle(self.read_path, **self.kwargs)
