+++ woodwork/serialize.py
@@ -1,0 +1,2 @@
+# This is included to prevent a breaking API change
+from woodwork.serializers.serializer_base import typing_info_to_dict
