+++ woodwork/serializers/arrow_serializer.py
@@ -1,0 +1,5 @@
+from woodwork.serializers import FeatherSerializer
+
+
+class ArrowSerializer(FeatherSerializer):
+    format = "arrow"
